Move core::task::local_data to core::local_data
