+++ src/libcore/local_data.rs
