+++ src/libcore/task/mod.rs
@@ -47,7 +47,6 @@ use unstable::finally::Finally;
 #[cfg(test)] use comm::SharedChan;
 
 mod local_data_priv;
-pub mod local_data;
 pub mod rt;
 pub mod spawn;
